Revert "Refactor page to server component for better performance"
