--- frontend/components/ToolSelector.tsx
@@ -1,86 +1,0 @@
-'use client';
-
-import { useState } from 'react';
-import FatLossCalculator from '@/components/tools/FatLossCalculator';
-import N26Analyzer from '@/components/tools/N26Analyzer';
-
-type Tool = 'fat-loss' | 'n26';
-
-export default function ToolSelector() {
-  const [selectedTool, setSelectedTool] = useState<Tool | null>(null);
-
-  if (selectedTool) {
-    return (
-      <div>
-        <button
-          onClick={() => setSelectedTool(null)}
-          className="mb-6 flex items-center text-gray-600 dark:text-gray-300 hover:text-gray-900 dark:hover:text-white transition-colors"
-        >
-          <svg className="w-5 h-5 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" />
-          </svg>
-          Back to Tools
-        </button>
-
-        {selectedTool === 'fat-loss' && <FatLossCalculator />}
-        {selectedTool === 'n26' && <N26Analyzer />}
-      </div>
-    );
-  }
-
-  return (
-    <div className="grid md:grid-cols-2 gap-8">
-      {/* Fat Loss Calculator Card */}
-      <button
-        onClick={() => setSelectedTool('fat-loss')}
-        className="p-8 bg-white dark:bg-gray-800 rounded-xl shadow-lg hover:shadow-xl transition-all duration-300 hover:scale-105 text-left"
-      >
-        <div className="flex items-center mb-4">
-          <div className="w-12 h-12 bg-green-100 dark:bg-green-900 rounded-lg flex items-center justify-center">
-            <svg className="w-6 h-6 text-green-600 dark:text-green-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 7h6m0 10v-3m-3 3h.01M9 17h.01M9 14h.01M12 14h.01M15 11h.01M12 11h.01M9 11h.01M7 21h10a2 2 0 002-2V5a2 2 0 00-2-2H7a2 2 0 00-2 2v14a2 2 0 002 2z" />
-            </svg>
-          </div>
-          <h2 className="text-2xl font-bold text-gray-900 dark:text-white ml-4">
-            Fat Loss Calculator
-          </h2>
-        </div>
-        <p className="text-gray-600 dark:text-gray-300">
-          Calculate the percentage of fat vs muscle loss based on your calorie deficit and weight loss.
-        </p>
-        <div className="mt-4 text-blue-600 dark:text-blue-400 font-semibold flex items-center">
-          Open Tool
-          <svg className="w-5 h-5 ml-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
-          </svg>
-        </div>
-      </button>
-
-      {/* N26 Analyzer Card */}
-      <button
-        onClick={() => setSelectedTool('n26')}
-        className="p-8 bg-white dark:bg-gray-800 rounded-xl shadow-lg hover:shadow-xl transition-all duration-300 hover:scale-105 text-left"
-      >
-        <div className="flex items-center mb-4">
-          <div className="w-12 h-12 bg-blue-100 dark:bg-blue-900 rounded-lg flex items-center justify-center">
-            <svg className="w-6 h-6 text-blue-600 dark:text-blue-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 002-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z" />
-            </svg>
-          </div>
-          <h2 className="text-2xl font-bold text-gray-900 dark:text-white ml-4">
-            N26 Transaction Analyzer
-          </h2>
-        </div>
-        <p className="text-gray-600 dark:text-gray-300">
-          Analyze your N26 bank transactions, view spending patterns, and get insights into your financial data.
-        </p>
-        <div className="mt-4 text-blue-600 dark:text-blue-400 font-semibold flex items-center">
-          Open Tool
-          <svg className="w-5 h-5 ml-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
-          </svg>
-        </div>
-      </button>
-    </div>
-  );
-}
